pyproject.toml
[project]
name = "langgraph-agent-openwebui-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "aiohttp>=3.11.18",
  "fastapi>=0.115.12",
  "langchain>=0.3.25",
  "langchain-community>=0.3.24",
  "langchain-ollama>=0.3.3",
  "langchain-openai>=0.3.17",
  "langgraph>=0.4.5",
  "pandas>=2.2.3",
  "passlib>=1.7.4",
  "psycopg2-binary>=2.9.10",
  "pydantic>=2.11.4",
  "pyjwt>=2.10.1",
  "python-dotenv>=1.1.0",
  "python-multipart>=0.0.20",
  "scipy>=1.15.3",
  "sqlalchemy>=2.0.41",
  "uvicorn>=0.34.2",
  "vllm>=0.8.5.post1",
]
